task: organize pages and components
