+++ src/App.tsx
@@ -4,14 +4,12 @@ import { ThemeProvider } from 'styled-components';
 import theme from './styles/theme';
 import GlobalStyle from './styles/global';
 
-import Header from './components/Header';
-import Gradient from './components/Gradient';
+import Dashboard from './pages/Dashboard';
 
 const App: React.FC = () => {
   return (
     <ThemeProvider theme={theme}>
-      <Header />
-      <Gradient />
+      <Dashboard />
       <GlobalStyle />
     </ThemeProvider>
   );
