+++ src/pages/Dashboard/Header/Gradient/styles.ts
